Just getting the basics rewritten.
