+++ Cheepicus12.font
@@ -1,1 +1,1 @@
-{"CellHeight":12,"CellPadding":0,"CellWidth":12,"FilePath":"Cheepicus_12x12.png","IsDefault":false,"Name":"Cheepicus12"}
+{"FilePath":"Cheepicus_12x12.png","GlyphHeight":12,"GlyphPadding":0,"GlyphWidth":12,"Name":"Cheepicus12","SolidGlyphIndex":219}
