--- EditorFont.font
@@ -1,1 +1,0 @@
-{"CellHeight":16,"CellPadding":1,"CellWidth":8,"FilePath":"EditorFont_IBM8x16.png","IsDefault":true,"Name":"IBM"}
