+++ FileLoaders/TextSurface.cs
@@ -1,10 +1,39 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
+using SadConsole.Consoles;
+using System.Linq;
 
 namespace SadConsoleEditor.FileLoaders
 {
-    class TextSurface
+    class TextSurface : IFileLoader
     {
+        public bool SupportsLoad { get { return true; } }
+
+        public bool SupportsSave { get { return true; } }
+
+        public string[] Extensions
+        {
+            get
+            {
+                return new string[] { "surface" };
+            }
+        }
+
+        public string FileTypeName
+        {
+            get
+            {
+                return "Single Surface";
+            }
+        }
+
+        public ITextSurfaceRendered Load(string file)
+        {
+            return SadConsole.Consoles.TextSurface.Load(file);
+        }
+
+        public void Save(ITextSurfaceRendered surface, string file)
+        {
+            ((SadConsole.Consoles.TextSurface)surface).Save(file);
+        }
     }
 }
