+++ SadConsoleEditor.Platforms/Windows/SadConsoleEditor.Windows.csproj
@@ -55,8 +55,8 @@
       <HintPath>..\..\packages\SadConsole.Controls.3.0.4\lib\net45\SadConsole.Controls.Windows.dll</HintPath>
       <Private>True</Private>
     </Reference>
-    <Reference Include="SadConsole.Core.Windows, Version=5.0.1.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\..\packages\SadConsole.Core.5.0.1\lib\net45\SadConsole.Core.Windows.dll</HintPath>
+    <Reference Include="SadConsole.Core.Windows, Version=5.0.3.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\..\packages\SadConsole.Core.5.0.3\lib\net45\SadConsole.Core.Windows.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="SadConsole.GameHelpers.Windows, Version=4.0.0.0, Culture=neutral, processorArchitecture=MSIL">
