+++ SadConsoleEditor.Platforms/Windows/packages.config
@@ -3,6 +3,6 @@
   <package id="ColorMine" version="1.1.3.0" targetFramework="net45" />
   <package id="MonoGame.Framework.WindowsDX" version="3.5.1.1679" targetFramework="net46" />
   <package id="SadConsole.Controls" version="3.0.4" targetFramework="net46" />
-  <package id="SadConsole.Core" version="5.0.1" targetFramework="net46" />
+  <package id="SadConsole.Core" version="5.0.3" targetFramework="net46" />
   <package id="SadConsole.GameHelpers" version="4.0.0" targetFramework="net46" />
 </packages>
