+++ SadConsoleEditor.projitems
@@ -11,66 +11,20 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)Consoles\BorderConsole.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Consoles\QuickSelectPane.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Consoles\ToolPane.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Controls\CharacterPicker.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Controls\ColorBar.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Controls\ColorPicker.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\ColorPresenter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Controls\FileDirectoryListbox.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Controls\NoCheckRadioButton.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)EditorConsoleManager.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Editors\AnimationEditor.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Editors\DrawingEditor.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Editors\IEditor.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Editors\SceneEditor.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)EntityBrush.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Editors\LayeredConsoleEditor.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FileLoaders\Entity.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FileLoaders\IFileLoader.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FileLoaders\Scene.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FileLoaders\TextFile.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)FileLoaders\TextSurface.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Game1.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)GameHelpers\GameObjectMeta.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HSLColor.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IEntityBrush.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\AnimationFramesPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\AnimationsPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\BoxTool.Panel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\CharacterPickPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\Circle.Panel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\CustomPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\EntityNamePanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\FilesPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\LayersPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\LinePanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\RecolorTool.Panel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\Scene\AnimationListPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\Scene\EntityManagementPanel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\SelectionTool.Panel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Panels\ToolsPanel.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Program.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Settings.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\BoxTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\Circle.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\DrawTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\EntityCenterTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\EntityCollisionBoxTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\FillTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\ITool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\LineTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\RecolorTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\SceneEntityMoveTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\SelectionTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Tools\TextTool.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Undo.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Windows\AnimationPreviewPopup.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Windows\AnimationSpeedPopup.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Windows\CharacterQuickSelectPopup.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Windows\ColorPickerPopup.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Windows\NewConsolePopup.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Windows\OtherColorsPopup.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Windows\RenamePopup.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Windows\ResizeSurfacePopup.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Windows\SelectFilePopup.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -80,7 +34,9 @@
     <Content Include="$(MSBuildThisFileDirectory)Cheepicus_12x12.png">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Content Include="$(MSBuildThisFileDirectory)EditorFont_IBM8x16.png" />
+    <Content Include="$(MSBuildThisFileDirectory)EditorFont_IBM8x16.png">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
     <Content Include="$(MSBuildThisFileDirectory)Icon.ico" />
     <Content Include="$(MSBuildThisFileDirectory)quickselect.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
@@ -96,6 +52,8 @@
     <None Include="$(MSBuildThisFileDirectory)Cheepicus12.font">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Include="$(MSBuildThisFileDirectory)EditorFont.font" />
+    <None Include="$(MSBuildThisFileDirectory)Editor.font">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 </Project>
